poc/requirements.txt
kafka-python==2.0.2
pymodbus==3.6.8
